Attempt move to React 16.3
