+++ src/Field.tsx
@@ -1,10 +1,8 @@
-import * as PropTypes from 'prop-types';
 import * as React from 'react';
-
-import { FormikProps } from './Formik';
-import { getIn, isPromise, isFunction, isEmptyChildren } from './utils';
 import warning from 'warning';
-import { GenericFieldHTMLAttributes } from './types';
+import { connect } from './connect';
+import { FormikProps, GenericFieldHTMLAttributes } from './types';
+import { getIn, isEmptyChildren, isFunction, isPromise } from './utils';
 
 /**
  * Note: These typings could be more restrictive, but then it would limit the
@@ -84,24 +82,10 @@ export type FieldAttributes = GenericFieldHTMLAttributes & FieldConfig;
  * Custom Field component for quickly hooking into Formik
  * context and wiring up forms.
  */
-
-export class Field<Props extends FieldAttributes = any> extends React.Component<
-  Props,
-  {}
-> {
-  static contextTypes = {
-    formik: PropTypes.object,
-  };
-
-  static propTypes = {
-    name: PropTypes.string.isRequired,
-    component: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
-    render: PropTypes.func,
-    children: PropTypes.oneOfType([PropTypes.func, PropTypes.node]),
-    validate: PropTypes.func,
-    innerRef: PropTypes.func,
-  };
-
+class FieldInner<
+  Props extends FieldAttributes = any,
+  Values = {}
+> extends React.Component<Props & { formik: FormikProps<Values> }, {}> {
   componentWillMount() {
     const { render, children, component } = this.props;
 
@@ -122,7 +106,7 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
   }
 
   handleChange = (e: React.ChangeEvent<any>) => {
-    const { handleChange, validateOnChange } = this.context.formik;
+    const { handleChange, validateOnChange } = this.props.formik;
     handleChange(e); // Call Formik's handleChange no matter what
     if (!!validateOnChange && !!this.props.validate) {
       this.runFieldValidations(e.target.value);
@@ -130,7 +114,7 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
   };
 
   handleBlur = (e: any) => {
-    const { handleBlur, validateOnBlur } = this.context.formik;
+    const { handleBlur, validateOnBlur } = this.props.formik;
     handleBlur(e); // Call Formik's handleBlur no matter what
     if (validateOnBlur && this.props.validate) {
       this.runFieldValidations(e.target.value);
@@ -138,14 +122,14 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
   };
 
   runFieldValidations = (value: any) => {
-    const { setFieldError } = this.context.formik;
+    const { setFieldError } = this.props.formik;
     const { name, validate } = this.props;
     // Call validate fn
     const maybePromise = (validate as any)(value);
     // Check if validate it returns a Promise
     if (isPromise(maybePromise)) {
       (maybePromise as Promise<any>).then(
-        () => setFieldError(name, undefined),
+        () => setFieldError(name, undefined as any),
         error => setFieldError(name, error)
       );
     } else {
@@ -161,10 +145,10 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
       render,
       children,
       component = 'input',
+      formik,
       ...props
-    } = this.props as FieldConfig;
+    } = this.props as FieldConfig & { formik: FormikProps<Values> };
 
-    const { formik } = this.context;
     const field = {
       value:
         props.type === 'radio' || props.type === 'checkbox'
@@ -201,3 +185,5 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
     });
   }
 }
+
+export const Field = connect<FieldAttributes, any>(FieldInner);
