+++ src/index.tsx
@@ -5,3 +5,4 @@ export * from './withFormik';
 export * from './FieldArray';
 export * from './utils';
 export * from './FastField';
+export * from './types';
