--- src/types.ts
@@ -1,23 +1,0 @@
-import * as React from 'react';
-
-export interface SharedRenderProps<T> {
-  /**
-   * Field component to render. Can either be a string like 'select' or a component.
-   */
-  component?: string | React.ComponentType<T | void>;
-
-  /**
-   * Render prop (works like React router's <Route render={props =>} />)
-   */
-  render?: ((props: T) => React.ReactNode);
-
-  /**
-   * Children render function <Field name>{props => ...}</Field>)
-   */
-  children?: ((props: T) => React.ReactNode);
-}
-
-export type GenericFieldHTMLAttributes =
-  | React.InputHTMLAttributes<HTMLInputElement>
-  | React.SelectHTMLAttributes<HTMLSelectElement>
-  | React.TextareaHTMLAttributes<HTMLTextAreaElement>;
